requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
six==1.15.0
termcolor==1.1.0
